pyproject.toml
[project]
name = "wndmngr-db"
version = "0.1.0"
dependencies = [
    "pyodbc>=5.2.0",
    "pandas>=2.3.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "pyjanitor>=0.31.0",
    "openpyxl>=3.1.5",
    "ipykernel>=6.30.1",
    "sqlmodel>=0.0.25",
    "invoke>=2.2.0",
]
